+++ finetuner/__init__.py
@@ -1,3 +1,9 @@
+import hubble
 from dotenv import load_dotenv
+
+
+def login():
+    hubble.login()
+
 
 load_dotenv()