+++ tests/constants.py
@@ -1,2 +1,1 @@
-RUN_CONFIG_PATH = 'tests/resources/run-config.json'
-USER_ID_FOR_TESTING = 1
+FINETUNER_LABEL = 'finetuner_label'