+++ CHANGELOG.md
@@ -34,6 +34,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Create run endpoints. ([#387](https://github.com/jina-ai/finetuner/pull/387))
 
+- Add Hubble authentication. ([#388](https://github.com/jina-ai/finetuner/pull/388))
+
 ### Removed
 
 - Delete all unnecessary files from the previous project. ([#384](https://github.com/jina-ai/finetuner/pull/384))